pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.4.2"
django-filter = "^24.2"
django = "^5.0.6"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
